refactor: fix unit test
