+++ server/coordinator/procedure/scatter_test.go
@@ -10,30 +10,14 @@ import (
 
 	"github.com/CeresDB/ceresdbproto/pkg/clusterpb"
 	"github.com/CeresDB/ceresdbproto/pkg/metaservicepb"
-	"github.com/CeresDB/ceresmeta/server/cluster"
-	"github.com/CeresDB/ceresmeta/server/coordinator/eventdispatch"
-	"github.com/CeresDB/ceresmeta/server/etcdutil"
 	"github.com/CeresDB/ceresmeta/server/id"
-	"github.com/CeresDB/ceresmeta/server/storage"
 	"github.com/stretchr/testify/require"
-	clientv3 "go.etcd.io/etcd/client/v3"
-)
-
-const (
-	nodeName0                = "node0"
-	nodeName1                = "node1"
-	testRootPath             = "/rootPath"
-	defaultIDAllocatorStep   = 20
-	clusterName              = "ceresdbCluster1"
-	defaultNodeCount         = 2
-	defaultReplicationFactor = 1
-	defaultShardTotal        = 8
 )
 
 func TestScatter(t *testing.T) {
 	re := require.New(t)
 	ctx := context.Background()
-	dispatch := eventdispatch.NewDispatchImpl()
+	dispatch := MockDispatch{}
 	cluster := newTestCluster(ctx, t)
 
 	nodeInfo1 := &metaservicepb.NodeInfo{
@@ -121,23 +105,4 @@ func TestAllocNodeShard(t *testing.T) {
 	re.Equal("node0", shardView[0].Node)
 	re.Equal("node0", shardView[1].Node)
 	re.Equal("node1", shardView[2].Node)
-}
-
-func newTestEtcdStorage(t *testing.T) (storage.Storage, clientv3.KV, etcdutil.CloseFn) {
-	_, client, closeSrv := etcdutil.PrepareEtcdServerAndClient(t)
-	storage := storage.NewStorageWithEtcdBackend(client, testRootPath, storage.Options{
-		MaxScanLimit: 100, MinScanLimit: 10,
-	})
-	return storage, client, closeSrv
-}
-
-func newTestCluster(ctx context.Context, t *testing.T) *cluster.Cluster {
-	re := require.New(t)
-	storage, kv, _ := newTestEtcdStorage(t)
-	manager, err := cluster.NewManagerImpl(storage, kv, testRootPath, defaultIDAllocatorStep)
-	re.NoError(err)
-
-	cluster, err := manager.CreateCluster(ctx, clusterName, defaultNodeCount, defaultReplicationFactor, defaultShardTotal)
-	re.NoError(err)
-	return cluster
 }
