+++ server/coordinator/procedure/transfer_leader_test.go
@@ -7,29 +7,13 @@ import (
 	"testing"
 
 	"github.com/CeresDB/ceresdbproto/pkg/clusterpb"
-	"github.com/CeresDB/ceresmeta/server/cluster"
-	"github.com/CeresDB/ceresmeta/server/coordinator/eventdispatch"
-	"github.com/CeresDB/ceresmeta/server/etcdutil"
-	"github.com/CeresDB/ceresmeta/server/storage"
 	"github.com/stretchr/testify/require"
-	clientv3 "go.etcd.io/etcd/client/v3"
-)
-
-const (
-	nodeName0                = "node0"
-	nodeName1                = "node1"
-	testRootPath             = "/rootPath"
-	defaultIDAllocatorStep   = 20
-	clusterName              = "ceresdbCluster1"
-	defaultNodeCount         = 2
-	defaultReplicationFactor = 1
-	defaultShardTotal        = 2
 )
 
 func TestTransferLeader(t *testing.T) {
 	re := require.New(t)
 	ctx := context.Background()
-	dispatch := eventdispatch.NewDispatchImpl()
+	dispatch := MockDispatch{}
 	cluster := newTestCluster(ctx, t)
 
 	// Initialize shard topology
@@ -73,23 +57,4 @@ func TestTransferLeader(t *testing.T) {
 			re.Equal(nodeName1, node)
 		}
 	}
-}
-
-func newTestEtcdStorage(t *testing.T) (storage.Storage, clientv3.KV, etcdutil.CloseFn) {
-	_, client, closeSrv := etcdutil.PrepareEtcdServerAndClient(t)
-	storage := storage.NewStorageWithEtcdBackend(client, testRootPath, storage.Options{
-		MaxScanLimit: 100, MinScanLimit: 10,
-	})
-	return storage, client, closeSrv
-}
-
-func newTestCluster(ctx context.Context, t *testing.T) *cluster.Cluster {
-	re := require.New(t)
-	storage, kv, _ := newTestEtcdStorage(t)
-	manager, err := cluster.NewManagerImpl(storage, kv, testRootPath, defaultIDAllocatorStep)
-	re.NoError(err)
-
-	cluster, err := manager.CreateCluster(ctx, clusterName, defaultNodeCount, defaultReplicationFactor, defaultShardTotal)
-	re.NoError(err)
-	return cluster
 }
